requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
tqdm>=4.65.0
markdown>=3.4.0
python-docx>=0.8.11
PyYAML>=6.0 